kafka/server: hoist group::stages into its own header
Move the template and its three aliases to namespace scope so code
outside `group` can use them. `group` keeps class-scope aliases, so
its call sites are unchanged.
